fix: support Lambda instrumentation for `contextManager: 'patch'`; refactor Lambda tests (#3305)
Changes in #3285 broke Lambda instrumentation if the deprecated
`contextManager:'patch' config var was used. This restores support for that, FWIW.

Also, refactoring so we can drop AgentMock and TransactionMock under 
test/lambda/... to simplify:
- roll promises.test.js into the main test file
- roll trace-context.test.js into the main lambda test file
- roll transaction.test.js into main lambda test file
- rename main lambda test file to lambda.test.js
- drop now unused MockAgent
