+++ docs/lambda.asciidoc
@@ -61,14 +61,14 @@ Your Lambda function invocations should be traced from now on.
 === Features
 
 The AWS Lambda instrumentation will report a transaction for all function invocations
-and trace any <<compatibility-frameworks,support modules>>. In addition, the
+and trace any <<compatibility-frameworks,supported modules>>. In addition, the
 created transactions will capture additional data for a number of Lambda
-trigger types -- API Gateway, SNS, SQS, S3 (with the trigger is a single event),
+trigger types -- API Gateway, SNS, SQS, S3 (when the trigger is a single event),
 and ELB.
 
 A transaction will be reported for Lambda invocations that fail due to a
 timeout, crash, `uncaughtException`, or `unhandledRejection`. (This requires
-APM agent v3.46.0 or later and
+APM agent v3.45.0 or later and
 https://www.elastic.co/guide/en/apm/lambda/current/aws-lambda-arch.html[Elastic's APM Lambda extension]
 version 1.4.0 or later.)
 
@@ -80,4 +80,5 @@ version 1.4.0 or later.)
 * System and custom metrics are not collected for Lambda functions. This is both because most of those are irrelevant
 and because the interval-based event sending model is not suitable for FaaS environments.
 * Lambda instrumentation does not currently work when the (deprecated) <<context-manager,`contextManager: 'patch'`>> configuration setting is used.
-* The APM agent does not yet support a Lambda handler module that uses ECMAScript modules (ESM). That means a your handler file name should end with ".js" (and not have `"type": "module"` in package.json if you have one) or end with ".cjs". A handler file that uses the ".mjs" suffix will not be instrumented by the APM agent.
+* The APM agent does not yet support a Lambda handler module that uses ECMAScript modules (ESM). That means your handler file name should end with ".js" (and not have `"type": "module"` in package.json if you have one) or end with ".cjs". A handler file that uses the ".mjs" suffix will not be instrumented by the APM agent.
+
