+++ lib/lambda.js
@@ -438,6 +438,7 @@ function setS3SingleData (trans, event, context, faasId, isColdStart) {
 
 function elasticApmAwsLambda (agent) {
   const log = agent.logger
+  const ins = agent._instrumentation
 
   /**
    * Register this transaction with the Lambda extension, if possible.  This
@@ -527,34 +528,43 @@ function elasticApmAwsLambda (agent) {
     })
   }
 
-  function wrapContext (trans, event, context, triggerType) {
+  function wrapContext (runContext, trans, event, context, triggerType) {
     shimmer.wrap(context, 'succeed', (origSucceed) => {
-      return function wrappedSucceed (result) {
-        endAndFlushTransaction(null, result, trans, event, context, triggerType, function () {
-          origSucceed(result)
-        })
-      }
+      return ins.bindFunctionToRunContext(
+        runContext,
+        function wrappedSucceed (result) {
+          endAndFlushTransaction(null, result, trans, event, context, triggerType, function () {
+            origSucceed(result)
+          })
+        }
+      )
     })
 
     shimmer.wrap(context, 'fail', (origFail) => {
-      return function wrappedFail (err) {
-        endAndFlushTransaction(err, null, trans, event, context, triggerType, function () {
-          origFail(err)
-        })
-      }
+      return ins.bindFunctionToRunContext(
+        runContext,
+        function wrappedFail (err) {
+          endAndFlushTransaction(err, null, trans, event, context, triggerType, function () {
+            origFail(err)
+          })
+        }
+      )
     })
 
     shimmer.wrap(context, 'done', (origDone) => {
-      return wrapLambdaCallback(trans, event, context, triggerType, origDone)
+      return wrapLambdaCallback(runContext, trans, event, context, triggerType, origDone)
     })
   }
 
-  function wrapLambdaCallback (trans, event, context, triggerType, callback) {
-    return function wrappedLambdaCallback (err, result) {
-      endAndFlushTransaction(err, result, trans, event, context, triggerType, () => {
-        callback(err, result)
-      })
-    }
+  function wrapLambdaCallback (runContext, trans, event, context, triggerType, callback) {
+    return ins.bindFunctionToRunContext(
+      runContext,
+      function wrappedLambdaCallback (err, result) {
+        endAndFlushTransaction(err, result, trans, event, context, triggerType, () => {
+          callback(err, result)
+        })
+      }
+    )
   }
 
   return function wrapLambdaHandler (type, fn) {
@@ -644,13 +654,14 @@ function elasticApmAwsLambda (agent) {
       // Note: Wrapping context needs to happen *before any `await` calls* in
       // this function, otherwise the Lambda Node.js Runtime will call the
       // *unwrapped* `context.{succeed,fail,done}()` methods.
-      wrapContext(trans, event, context, triggerType)
-      const wrappedCallback = wrapLambdaCallback(trans, event, context, triggerType, callback)
+      const transRunContext = ins.currRunContext()
+      wrapContext(transRunContext, trans, event, context, triggerType)
+      const wrappedCallback = wrapLambdaCallback(transRunContext, trans, event, context, triggerType, callback)
 
       await registerTransaction(trans, context.awsRequestId)
 
       try {
-        var retval = fn.call(this, event, context, wrappedCallback)
+        const retval = ins.withRunContext(transRunContext, fn, this, event, context, wrappedCallback)
         if (retval instanceof Promise) {
           return retval
         } else {
