+++ test/lambda/github-issue-2709.test.js
@@ -6,11 +6,6 @@
 
 'use strict'
 
-if (process.env.ELASTIC_APM_CONTEXT_MANAGER === 'patch') {
-  console.log('# SKIP Lambda instrumentation currently does not work with contextManager="patch"')
-  process.exit()
-}
-
 const tape = require('tape')
 const path = require('path')
