+++ test/lambda/metadata.test.js
@@ -8,11 +8,6 @@
 
 // Test that a lambda invocation sends the appropriate "metadata" to APM server.
 
-if (process.env.ELASTIC_APM_CONTEXT_MANAGER === 'patch') {
-  console.log('# SKIP Lambda instrumentation currently does not work with contextManager="patch"')
-  process.exit()
-}
-
 const lambdaLocal = require('lambda-local')
 const tape = require('tape')
