+++ test/lambda/trans-setDefaultName.test.js
@@ -10,10 +10,7 @@
 // works after the Lambda instrumentation has set the transaction name.
 
 const semver = require('semver')
-if (process.env.ELASTIC_APM_CONTEXT_MANAGER === 'patch') {
-  console.log('# SKIP Lambda instrumentation currently does not work with contextManager="patch"')
-  process.exit()
-} else if (semver.satisfies(process.version, '>=10.0.0 <10.4.0')) {
+if (semver.satisfies(process.version, '>=10.0.0 <10.4.0')) {
   // This isn't considered an issue because AWS Lambda doesn't support a node
   // v10 runtime, and node v10 is EOL.
   console.log(`# SKIP async context propagation currently does not work to a Lambda handler with node ${process.version}`)
